Strugling with RecyclerViews
Having Trouble Accessing the Classes I need to when creating the data objects for each case.

Note: For each of the menus, there are requirements.

Pending Cases: Cases where you disagree on the original bet's outcome

My Cases: You Chose to to accept or decline and then arbitrate

Rated Cases: Cases that are complete that you had no involvment in.